requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
requests==2.28.2
python-dotenv==1.0.0